Rendering OpenVDB volumes in hydra storm:
Adding GlfVdbTexture for creating a 3d OpenGL texture from an OpenVDB file. It is always using the first grid in the OpenVDB file. The transform associated with the OpenVDB grid and the bounding box of the tree can be accessed as GlfVdbTexture::GetBoundingBox (but is not used yet when rendering the volume in hdSt). Implementation details are in GlfVdbTextureData.

Adding HdStField that creates a HdStFieldResource using GlfVdbTexture. Updating HdStVolume to sample from the 3d texture.

Some work remains for the CMake build: When CMake detects that OpenVDB is present, it needs to add -DPXR_OPENVDB_SUPPORT_ENABLED as compiler flag and glf/vdbTextureData.h/cpp to the files to be compiled.

(Internal change: 1995667)
(Internal change: 1995675)
(Internal change: 1995676)
(Internal change: 1996490)