Fix UsdShade shader material invalidations in UsdImaging
For full materials (those loaded with the materialAdapter.cpp) in HdPrman, we were registering (calling AddHdPrimInfo) the material prim and all shader nodes when we call "Populate" in the materialAdapter. However, this population only happens when the Gprim that is using that material is populated either during initial population or gprim invalidation like changing the binding. This means that changes localized inside the material network, like adding new shader nodes, might not be correctly tracked by UsdImaging (see test case).

This fix makes material population only register the material itself (not the shader nodes) in UsdImaging. When there are shader node invalidations we trace back the material and properly invalidate it.

Fixes #617

(Internal change: 1996060)