Update comment to state render buffer map is for read only.
Writting this way require more careful communication with the backend.
For example, to write to a GPU buffer we would need to persistently map the buffer, or have a upload-to-gpu in unmap.
We would need to know when new data is uploaded.

For now we should be clear this is only supporting read.

(Internal change: 1996362)