Volumetric rendering in storm:
Adding raymarching step sizes as storm render settings. A new subclass of HdStSurfaceShader will consume them from the render delegate and bind them so that they can be used in the volume raymarching code. One of the step size settings will affect the raymarching of the camera ray. The other one will affect the raymarching from a point in the volume to a light source to compute the in scatter amount.

(Internal change: 1997166)