Adding GetValueProducingAttribute to UsdShadeInput
This is a method to compute where a shading input will get its value from, which can be either the output of an actual shading node, something that will be computed at render time via a connection, or it is a constant value that is the value of an input attribute.

Adds testUsdShadeGetValueProducingAttribute to execise this new method.

* Added TRACE_SCOPE to UsdShadeConnectableAPI::GetConnectedSource() for better tracking
* Expose missing UsdShadeOutput::IsOutput to python
* Adding UsdShadeAttributeType::Invalid to be able to return an indetermined type from GetValueProducingAttribute in the case where no valid attribute was found

Notes:
* The implementation does not attempt to be ultra robust to potentially invalid networks. It could get stuck in an infinite loop if there is an infinite loop in the shading network
* It also does not detect and or reject certain connections that are invalid in the UsdShade scene. Certain self reflective connections are possible in general Usd but are not legit in a shading network. This implementation does not try to enforce this.

(Internal change: 2003796)