riley never gets destroyed in hdxPrman
iley wasn't actually getting destroyed in ~hdxPrman.  So if hdxPrman was recreated, the contructor would then attempt to create a new riley, but it has some static stuff that needs to get cleaned up in the destructor.

Also, HdxPrmanFramebuffer wasn't resetting the static nextID to 0 when destroyed, so on a subsequent attempt to call HydraDspyImageOpen, it would fail with an unknown buffer ID.

(Internal change: 2004374)