Making UsdShadeInput::GetValueProducingAttribute more robust
We're adding cycle detection to prevent an infinite loop in case of cycles in the attribute connections. Using a TfSmallVector to keep the cost of this check as small as possible.

Changed the logic in _GetValueProducingAttributeRecursive a bit to track the type of the last valid attribute, even if we have not found an attribute that carries an authored value. This was necessary to only pick up a value if we had not encountered an issue (invalid input or output, or a cycle).

If an issue was detected, it was possible that a previous recursion level would find a valid value and return it, thus masking the error. The new scheme will correctly return the error and only a legit attribute chain will resolve to the last attribute with a valid authored value.

(Internal change: 2004451)