Add support for UsdShadeNodeGraph to UsdImaging
The previous version of the material processing was oblivious to NodeGraph prims and ignored them completely. This fix adds support for that.

This also optimizes the _WalkGraph method in two ways:
* Use an SdfPathSet to identify previously visited nodes, instead of a linear search for list of nodes found so far. This should improve the scalability with larger networks
* Combined the loops that would find upstream nodes and visit them and then create the connections or extract values. Now there is a single loop, which should reduce the overhead

The new loop makes use of a smart helper method called GetValueProducingAttribute, which has been added to UsdShade, which will determine where an input of a shading node will come from. Only this method is aware of NodeGraph prims and steps right through them along connections. This method also encodes our public interface resolution logic, which will follow connections to input attributes and take the last valid value along that chain.

Removed the bogus fallback to 'PbsNetworkMaterialStandIn_2' as a fallback shader.

(Internal change: 2004737)