RenderMan Light Customizations
Translate RenderMan-only parameters that aren't exposed in the usdLux schema.
RenderMan-specific parameters are prefixed with the namespace "ri:light" to be
consistent with the rman-specific attributes ("ri:attributes").

Since there is no way to iterate over light parameters, so we need to map each directly.
At some point, this code should be cleaned up be a map of hydra token -> rman ustring.

(Internal change: 2004988)