Fix various GL errors when displaying an Aov with HdPrman.
When HdPrman runs, ColorizeTask takes the pixels and, via Compositor, puts them on the framebuffer.
Then color correction is run on the framebuffer.
(This is wrong, since we are now color correcting 8bit pixels, but that is for another day)

ColorCorrection checked _aovBuffer to decide if it should color correct an aov or the framebuffer.

When toggling back and forth between different AOVs, the AOV name is set by TaskController in color correction.
This means _aovBuffer would suddenly be found, even though we don't want to color correct the aov, we wanted the framebuffer.

It is safer to check for _aovTexture, which does a more robust check for find the AOV texture.

But there is plenty of room for improving this entire setup and make it more robuts.
I've also disabled any color correction on non-color AOVs for now  (TBD).

(Internal change: 2007775)