There was a crash in UsdImagingGLEngine::SetRendererPlugin if _rendererPlugin->CreateRenderDelegate() returned a nullptr.  There was no check for whether the resulting delegate pointer was valid so a nullptr ended up getting dereferenced in UsdImagingDelegate constructor.
This change adds a check for whether CreateRenderDelegate returned null and returns false from SetRendererPlugin, so the caller can do something reasonable like switch to a fallback delegate.  The call to CreateRenderDelegate has also been moved earlier in the method, to happen before _DeleteHydraResources, so that UsdImagingGLEngine will be left in a reasonable state if CreateRenderDelegate fails.

(Internal change: 2008070)