[HdxPrman] Fix selection highlighting.
Hydra assigns prim IDs with "-1" reserved as the no-prim marker.  In HdPrman, we put prim ID into "identifier:id", so that we can get per-pixel prim ID output.  However, the default value of "identifier:id" (used in case of a ray miss) is 0.

This meant that both object "0" (in hydra terms) and the background in HdPrman were outputting primId of 0, and were getting selected together.

The fix is to map "identifier:id" as (hydra-id + 1).  We only make this change to primId, but in the framebuffer code if a pixel has "-1" for primId we assign "-1" for all other IDs as well.

(Internal change: 2008073)