[usdImaging] Fix up data ownership of instance indices.
DirtyInstanceIndex no longer applies to rprims.  DirtyInstanceIndex is now set on instancers and tells them they need to re-fetch their index buffer, relative to one of their prototypes.

On the hydra side:
- This involved fixing up a lot of copy & pasted initial dirty bits, where folks would set DirtyInstanceIndex on rprims to pick up instancer data.  Now, we add DirtyInstancer instead, which is the bit for "rprim has an instancer dependency that's been updated".

On the USD side:
- Since DirtyInstanceIndex is handled by instancer adapters, for the instancer prim in question, we can compute index data on the fly.  We no longer need to:
1.) For prototypes, use adapter hijacking to insert ourselves into the prototype UpdateForTime call.
2.) For nested instancers, retrieve the parent instancer's adapter and use special API to get its index buffer.

Both of these made the code much harder to follow, and required mutexes to guard against the multiple children asking for parent index data at the same time.  Getting rid of said mutexes and the adapter hijacking model are an end goal of this project.

As an added bonus, when index data changes on an adapter (because of population changes or adapter updates), it's easy to mark the correct dirty bit.  Before, our approach was to invalidate the instancer and everything downstream (and possibly upstream) and hope for the best.

While the responsibility for index computation has changed, and dirty bit semantics have changed, the prototype index data is still stored per-prototype in the value cache, since the prototype path is the unique key to that buffer.

(Internal change: 2014494)