Remove the context switching check in the Color Correction and Color Channel tasks.
Calls like  context->IsCurrent() and GlfGLContext::GetCurrentGLContext() can be slow if more than one context is registered in GlfGLContextRegistry, as the GlfGLContextRegistry::GetCurrent() function will iterate through each context, making each one of them current in the loop.

For now we can assume that there is no context switching between the creation of the render pass and the execution of these tasks on each frame, so this code is being removed.

(Internal change: 2020716)