[usdGeom] Add XformCommonAPI to schema.usda as a non-applied API schema and codegen the class skeleton
This is a mostly straightforward adaptation of the existing
UsdGeomXformCommonAPI to fit the codegen template, with
the notable exception that the _xformable and _xformOps
members are not initialized upon construction.

Usage of cached _xformable is replaced with
UsdGeomXformable(GetPrim()) where necessary.

Usage of cached _xformOps has been replaced with either
a direct retrieval of the ordered xform ops OR caching the
xform ops when computing the op indices within a
non-const function.

(Internal change: 2021708)