Adding option to build OpenVDB (off by default) to build_usd.py, the script that downloads USD dependencies and compiles them and USD.
This change also changes the boost version on Linux from 1.55.0 to 1.61.0.

The change also checks for CMake to be version 3.3 or later, but only if OpenVDB support is requested. The OpenVDB version downloaded is 6.1.0. When requesting OpenVDB support, we also download and compile blosc (OpenVDB's preferred compression), version 1.17.0.

(Internal change: 2027186)