The pick task is continuously running in UsdView because as you move the mouse it uses the pick task to see what you are hovering over. It then displays a pop-up window with info of the hovered item.
When you move the mouse to the 'X' button to close UsdView the pick tasks may be one of the last things that gets executed.
It does a DoneCurrent, which sets the gl context to null.

Then we try to cleanup the gl resources, such as deleting framebuffers and textures.
Those will fail because the gl context is null.

Solution is to not call DoneCurrent during pick. It isn't really needed.

(Internal change: 2027189)