[usdImaging] Fix instance adapter selection highlighting performance issues.
(Possibly some correctness too).  The PrimAdapter PopulateSelection API previously only took "cachePath", a pointer to the hydra prim being selected.  For native instancing, you also need the path to the USD instance, so you know which instance to highlight.  We had some sketchy code in UsdImagingDelegate that would test whether the highlighted prim was an instance, and if so, replace "cachePath" with the USD path of the instance.

This loses the information of which hydra instancer is affected, so the native instance adapter was looping through all native instances in the scene.

By adding "usdPrim" to the PrimAdapter PopulateSelection signature, we can clean up delegate PopulateSelection and choose the right native instancer to call PopulateSelection on.

(Internal change: 2028242)