There is a sanctioned way to have non-identifier tokens anywhere in a schema.usda file.  For "allowedTokens", this is handled for you transparently.  Unfortunately, we for some reason decided way back to force you to be more explicit for libraryTokens and schemaTokens, adding a separate "value" field to the specified dictionary.  This procedure is documented, and we do not intend to deprecate it.
A recent change attempted to bypass this requirement by adding
a TfMakeIdentifier call into the processing loop, after an illegal
libraryToken happily passed through usdGenSchema only to cause
the subsequent compilation to fail.

To avoid having two ways of doing the same thing, we now remove
the TfMakeIdentifier call, and instead issue an informative error if
a libraryToken or schemaToken is non-identifier.

(Internal change: 2028356)