UsdMtlxParserPlugin::Parse() now finds implementation files correctly.
Previously the code assumed that implementation file relative paths
(e.g. paths to an OSL or GLSL file) were relative to the mtlx file
with the implementation node.  This is not correct.  They're relative
to the MaterialX library, which is identified by the path to its
top level directory.  This path is stored neither anywhere in the
MaterialX document nor in the NdrNodeDiscoveryResult.

We now search the list of MaterialX library paths, joining the
relative path to each directory in turn until we find a file that
exists and returning the path to it.

UsdMtlxStandardLibraryPaths() now returns a const reference instead
of a copy.  There's no reason it should not have been doing so and
now we iterate over the vector for every implementation relative
file path so the copy would unnecessarily slow that down.

(Internal change: 2039915)