[Tf] Changes for Python 3 support
Adds support for building Tf in Python 2 or 3.

These are initial steps for adding Python 3 support to USD,
with more to come.

This commit is based off work done by NVIDIA and contributed
in PR #928 as well as additional changes contributed by
SideFX. Thanks to both teams for their generous support!

(Internal change: 2042550)