Refactoring of how Usd_PrimData accesses prim type information and prim definitions from the schema registry in preparation for adding applied schemas to the prim type.
Part 1 is the addition of the new class UsdPrimDefinition that serves as a wrapper around the mapping of property names to property spec paths for a registered schema prim type. The prim definitions for each registered schema type are created when the schema registry is instantiated and provide the API for accessing property specs and property fields like the schema registry used to. They are accessed by name from the schema registry. The purpose of this abstraction is that we can later add the ability to compose a UsdPrimDefinition from multiple existing definitions based on a combination of prim type and applied API schemas, provided a custom mapping to properties provided from multiple sources.

Part 2 is a new private Usd_PrimTypeInfo class has been added for use by the Usd_PrimData as a way of storing the prim's type instead of as just a type name token. Each unique type is stored in a private static type registry and the Usd_PrimData stores a pointer to its type info. The Usd_PrimTypeInfo provides a function to get and cache a pointer to the UsdPrimDefinition for its type from the schema registry. The purpose of this new type info is so that we can later compose and add the list of applied API schemas to the type info for the prim caching each unique combination as its own type. The prim data stores a pointer to the correct compound type. When the prim needs to access the UsdPrimDefinition for a compound type, it will be able to build and cache the new UsdPrimDefinition in its type info without having to mutate the schema registry.

There are also some cleanups to the SchemaRegistry itself. Deprecating the GetSchema...Spec functions to be removed later in favor of clients using the prim definition. Making the remaining static function that instantiate the singleton not static.

(Internal change: 2042939)
(Internal change: 2043117)
(Internal change: 2043153)