Route ri:attributes:* to either the geometry master as a primvar or to the geometry instance as an attribute
Adds _IsMasterAttribute, which contains an explicit list of ri:attributes:* attributes that should authored on the geometry master as primvars, instead of becoming attributes on the geometry instance.

The routing in _Convert has been adjusted to consult this function and put the ri:attributes: into the right place.

(Internal change: 2043813)