Add an 'existOk' parameter to TfMakeDirs.
Currently, TfMakeDirs always returns false when asked to create a directory
that already exists. Because of this, the typical pattern for creating
directories looks like this:

    if (not (TfIsDir(dirPath) or TfMakeDirs(dirPath)) {
        // ... error
    }

However, this is prone to thread races, where multiple threads see
TfIsDir as false, one thread creates the directory, and the others
get a false return from TfMakeDirs and report the error, even
though the directory does now exist.

With this change, the typical pattern can be simplified:

    if (not TfMakeDirs(dirPath, /* mode */ -1, /* existOk */ true)) {
        // failure due to something other than preexisting directory.
    }

(Internal change: 2047007)