hdPrman: avoid std::move in return
'return std::move(local);' prevents the compiler from applying copy
elision optimizations.  While copy elision is not guaranteed to be
applied, if the compiler is unable to apply copy elision, it is
required[1] to select a constructor as if 'local' was an r-value, making
the call to std::move unneeded in any case.

[1] Except if 'local' is an argument to the function.

(Internal change: 2047392)