Support HdAov clearValue in HdPrman.
In framebuffer we check if we have a ray-miss (no primId) and use clearValue when there is a miss.
If there is a hit, blend hit-color with background based on alpha.

In RenderPass we check if the clear color has changed since last time.
This allows application like UsdView to change background color.

When the camera rotates, we need to clear depth since we will no longer be doing the composite via a gl-fullscreen triangle in Hdx in a future changelist.

(Internal change: 2047933)