[usdImaging] Selection highlighting overhaul.
The existing selection highlighting code had some limiting assumptions:

1.) Hydra rprims would only depend on USD gprims.

This was broken when we added a dependency (via the new dependency map) between gprims and materials.  Before this CL, if you select a material in the browser it will highlight bound gprims.  While this is a cool feature, it worked because of a bug and this has been addressed.

2.) Instancing was only single-level.

The old code for selection highlighting in instancers would take the instancer name (e.g. "/Instancer"), and do a hydra search for all rprims of the name ("/Instancer.*"), applying a highlight to them.  This doesn't work for nested instancing.  The correct thing to do is to indirect through child prim adapters.  The resulting code is a bit more verbose/complicated, but it works for nested instancing cases and no longer calls HdRenderIndex::GetRprimSubtree.

3.) Selection of an instance proxy prim was converted to a selection of the containing top level instance.

This meant that if you had multiple prims in a master, you couldn't meaningfully select only one of those prims.  This breaks the ideal of selection highlighting not particularly caring whether a prim is an instance proxy or a real prim.  Now, we preserve the actual instance proxy selection and pass that through to adapters.

With these changes, selections in the browser should show up as expected in the viewport.  This has been tested against the diversity of nested instancer cases in usdImagingGL.

Also, as some bonus cleanup:
- Adapter::PopulateSelection is now const.
- Adapter::_GetRprimSubtree has been deleted.

Fixes #713

(Internal change: 2048544)