Some clean-ups in HdStVolume: de-duplicating field accessors is achieved by recording field names first before looping over them instead of a guard preventing second instantiation.
Note that there is a tiny change in behavior with respect whether to use the authored value or computing the bounding box volume. Whereas previously, the computed value was used whenever there was a field reader, it is now used whenever there is a field reader whose field is actually connecting to a field asset prim.

(Internal change: 2049336)