Bump minimum required CMake version to 3.12
USD had previously required either 2.8.8, 3.1.0,
or 3.1.1 depending on a variety of factors. Support
for 2.x was deprecated in 19.11. 3.12 was picked as
the minimum requirement because it's relatively recent
(released in 2018) and it's the first release that
fixes target_link_libraries() to support object
libraries. This may let us remove a bunch of code
in the build system that worked around this limitation
for monolithic builds.

(Internal change: 2049628)