Bump Alembic version in build_usd.py to 1.7.10
Change 2049435 caused test failures on Windows when
using version 1.7.1. These failures went away starting
in version 1.7.6, indicating a bug that was fixed in
Alembic itself. Instead of just bumping to 1.7.6, we
go to 1.7.10 since that's the version used internally.

This version bump also allows us to remove a patch to
Alembic that's no longer necessary.

(Internal change: 2050162)