[tf] Fix testTfPathUtils for Python 3 on Windows
In python 2 the os module has no symlink function in windows. In python
3 there is a symlink function, but it will fail for non-administrator
users. Also, it will succeed in "developer mode" on some versions of
Windows 10. To avoid spurious failures or requiring administrator
access, we skip the symlink-specific portion of the test if creating
a symlink fails on Windows.

(Internal change: 2051465)
(Internal change: 2051479)