Remove HgiImmediateCommandBuffer and replace it with deferred command recording for OpenGL.
The idea is that we want to be able to use 'encoders' with a deferred-recording mindset.
This works great for modern API's like Dx12, Metal and Vulkan who support command buffers.

In OpenGL, the command execution is immediate. To make it deferred we record a lambda fn for each Hgi function we wish to execute and only emit the opengl calls when encoder::Commit is called.

Since HdSt is not yet using Hgi calls everywhere, we temporarily introduce 'InsertFunctionOp' on the encoders to allow HdSt to insert additional functions into the encoder to ensure the correct ordering of opengl calls.

HgiGLDevice was introduces to carry shared state, such as the framebuffer cache.
(Most other gfx API's will already require a 'device' object, so this is inline with Metal/Vulkan/Dx12)

(Internal change: 2051957)
(Internal change: 2052183)