Storm: moving determination of whether the material uses any ptex from HdSt_MaterialBufferSourceAndTextureHelper to HdStMaterial.
Note that the new code is using the texture type that is coming from the material network to make the determination whereas the old one looks at the type coming from the texture resource.

Those should give the result for all sane cases. Corner cases where it might now report there is a ptex (and quadrangulate) is if the texture resource cannot be found.

(Internal change: 2052127)