Fix Python detection when running build_usd.py with mayapy
The various sysconfig flags that build_usd.py relies on have
are set differently in mayapy so the code that tries to
determine the Python include and library paths failed in
various ways on different platforms. This change adds special
handling to accommodate this.

This fixes a regression that was introduced as part of the
Python 3 work in 20.05. However, we do plan to remove all
Maya-specific handling in build_usd.py in the future in
favor of code or instructions that live with the Maya
plugin itself.

(Internal change: 2055230)