Storm volumes and material params:
Follow-up to change 2055702 which simplified using a field texture for the client. In particular, the volume client no longer has to transform material params coming from field readers in a material. This means that we can actually remove one (enum) type of material params.

Renaming HdStMaterialParam::ParamTypePrimvar to ParamTypePrimvarRedirect for consistency.

(Internal change: 2056004)