Convert most tasks to use Hgi.
*** This change is a breaking change for OpenGL on MacOS which will be fixed in an upcoming change ***

- Introduce AovInputTask which takes the AOV (HdRenderBuffer) pixel data and, if needed, uploads it to the GPU as a HgiTexture.
  It places these HgiTextures onto the TaskContext so all following tasks can easily operate on them.
  AovInputTask also resolves the incoming AOV, so we no longer need AovResolveTask.

- Convert ColorCorrectionTask, PresentTask and ColorizeSelectionTask to use Hgi.

- Convert FullscreenShader (compositor) to use Hgi.

- Upgrade fullscreen.glslfx to modern glsl.
- Upgrade outline.glslfx to modern glsl and use a SSBO for parameters.
- Upgrade ColorCorrection.glslfx to modern glsl (and work-around the glsl 130 limit of OCIO)

- Update task controller to mostly be following the same tasks between Storm and other renderers like Renderman / Embree.

- We can now use the AOV resolution to determine the backbuffer size, so we no longer need to pass in 'renderResolution'.
- We no longer need the glViewport capture / override in HdSt renderPass.

- Embree UsdImagingGL update to black background:
This changelists removes the 'fullscreen triangle blend' step we did during Colorize to blend the pixels from Embree onto the gl framebuffer.
(We now directly copy the CPU pixels into a GPU texture)
While the gl framebuffer for the GL tests is cleared to range, the AOV clearColor given to Embree is black.
This is because a default HdxSetupTask is created, which creates default aov bindings with a black clear value.
Storm render directly to the gl framebuffer, so it has a orange background.
While Embree renders to its AOV which has a black background.

(Internal change: 2056036)
(Internal change: 2056161)