Storm texture system: udim support.
Udim textures are communicated from the scene delegate to the render delegate like all other texture types are communicated in the new texture system by an asset-valued parameter on a texture node in a material network. For a udim texture, the resolved part of the SdfAssetPath will be a resolved file path such as /filePath/myImage.<UDIM>.exr.

Thus, this change also involves a change to the usd scene delegate that finds asset-valued attributes in a material network and finds the first udim tile to use its resolved path as basis for the resolved file path.

The texture system (in the render delegate) is expanding the <UDIM> the resolved file path by different integers to find all udim tiles.

This change also removes the ptex sampler object trying to destruct texture handles (which might not work for the reasons indicated in the other sampler object d'tors.

(Internal change: 2058293)