Adding code so that scene delegates that rely on HdSceneDelegate::GetTextureResource/ID can continue to work while we switch to the new storm texture system.
The trigger whether to create a texture using the storm texture system or the HdSceneDelegate::GetTextureResource/ID is type of the file attribute on a texture node in the HdMaterialNetwork. If it is SdfAssetPath or std::string, the storm texture system is used. Otherwise, HdSceneDelegate::GetTextureResource/ID is used (e.g., PhD is using an SdfPath as file attribute).

The necessary data to invoke HdSceneDelegate::GetTextureResource/ID are funneled through the HdStMaterialNetwork::TextureDescriptor's (instead of the HdStMaterialParam).

(Internal change: 2058512)