Storm: changing how fallback values for uv and field textures are managed. The generated GLSL code has a branch using the fallback value if a texture is not valid (e.g., the file does not exist). The branch checks either the value of NAME_valid (in the shader bar) or whether the bindless texture handle is non-null. The fallback value is taken from NAME_fallback (in the shader bar).
Since this mechanism is consulting the shader bar, it is only used for bindings coming from material params from the material shader. This is flagged in HdSt_ResourceBinder::shaderParameterBinding::handleTextureFallbackValue.

On the client side, this means that HdStMaterial and HdStVolume needs to add buffer sources for NAME_fallback and HdSt_TextureBinder and HdSt_MaterialBufferSourceAndTextureHelper to add buffer sources for NAME_valid (if not using bindless textures).
The storm texture system communicates an invalid texture through HdStTextureObject::IsValid() which is consumed by HdSt_TextureBinder and the HdStSamplerObject's.

(Internal change: 2059483)