Remove detection of flex and bison from sdfGenAsciiParsers
sdfGenAsciiParsers previously tried to extract the paths to
flex and bison from the build system. However, the CMake
codepath didn't work as intended and the scons codepath
was specific to Pixar's internal build system. This was
also a fairly large amount of code -- roughly 20% of the
script.

sdfGenAsciiParsers now requires clients to explicitly
pass in the paths to flex and bison at the command line.
This was already being done in sdf's CMakeLists.txt for
during parser validation.

(Internal change: 2066793)