Bug fix: We were not setting any attachment descriptor for ColorSelectionTask. This causes a fatal error in HgiMetalPipeline.
Decided to take this bug to cleanup FullscreenShader a bit more so the user doesn't provide an entire pipeline descriptor of which we only took SOME settings). Instead the user provides the depth state.

This change also includes a few extra Metal cleanups.
A few HgiMetal headers need to be public since they are used by HgiInterop

(Internal change: 2067503)