Update hydra selection during task sync instead of prepare - Rename HdxSelectionTracker::Prepare to UpdateSelection - Call UpdateSelection during HdxSelectionTask::Sync instead of Prepare.
This is preparatory work to move Storm's selection highlighting to a stand alone task to allow customization of the selected look (e.g. showing occluded selected objects).
This render pass needs to be created during Sync to comply with the existing HdRenderIndex::SyncAll algorithm (so the collection's repr is sync'd). The HdSelection was being updated during Prepare, making it late in the Hydra execution model to create the collection.

(Internal change: 2068427)