Fix bug where ancestral value clips were ignored
For example, if value clips were specified on prim /A
and contained values for attributes on /A/B, but then
more clips were specified on prim /A/B/C, the values
for /A/B's attributes would be ignored. This was due
to ancestral clips not being copied to intermediate
entries in the SdfPathTable in cases like this.

(Internal change: 2068760)