[hd] make render index responsible for computing dirty rprim ID lists
With the removal of task-based collection include/exclude path filtering,
HdDirtyLists were generating their lists of dirty rprim IDs by looking
through every rprim in the render index. When the number of tasks/render
passes/dirty lists grew large, this resulted in significant overhead and
lots of duplication of work.

As an interim solution, this change makes the render index itself take care
of generating the complete list of dirty rprim IDs when requested by the
HdDirtyList. During SyncAll(), the first HdDirtyList to request the list of
dirty IDs for a given HdDirtyBits mask triggers the render index to produce
that list and cache it in a map. Subsequent requests reuse the cached list.
At the end of SyncAll(), the map is cleared in preparation for the next
sync.

(Internal change: 2069152)