Fix "Value" display for attributes in Layer Stack tab.
The relevant function name 'GetValueAtFrame' was pretty misleading,
as was its tricky implementation (returning different kinds of output
for different inputs).  So we combined its functionality with
GetShortStringForValue(), which is always called after it, to always
and unambiguously return both a value and a value string,
changing its name to indicate what it does.

We actually return a value-producing function, so that even
SdfAttributeSpecs can produce a "correct" value of "all
timeSampled values" if asked, which is expensive, and no
current GUI element ever asks for the full value from
AttributeSpecs.

FInally, cleanup use of the 'timeSamples' field, which we
do our best to keep hidden from python as an implementation
detail, and furthermore was being consulted unnecessarily.

(Internal change: 2070893)