Add tessellation and geometry stages to Hgi.
This helps transition HdStGlslProgram to Hgi.

For geometry shader, we wish to try and remove them from Storm completely and replace them with alternatives (compute shaders or doing additional work in tessellation stage).

For tessellation stages, we can support them as-is in GL and Vulkan.
Metal will have its own alternative to tessellation (See Metal implementation of OpenSubDiv) so we can probably ignore the stages coming in for the most part.

(Internal change: 2074704)