[usdImaging] When calling SampleTransform/SamplePrimvar, clear the transform/primvar out of the value cache.
Prim UpdateForTime() stashes primvars and transforms in the value cache.  Get() and GetTransform() will clear them from the value cache and pass them along to hydra, but if the render delegate chooses the "Sample" api instead, the value cache entries never get cleared out.

This change clears the value cache entries in SampleTransform/SamplePrimvar, on the theory that you'll never call "SamplePrimvar" and "Get" on the same primvar.  It should improve hdPrman memory usage considerably.

This change also necessitated a cleanup in HdPrmanInstancer, to not call Get() on the transform primvars (which we're already calling SamplePrimvar on).

(Internal change: 2075205)