Added support for arbitrary numbers of AOVs in hdPrman.
hdPrman had a hardcoded list of outputs that it would always create in Riley, ignoring the list of "RenderVars" passed down from USD via Hydra
via the HdRenderPassAovBindingVector that is accessible in HdxPrman_RenderPass::_Execute.
The creation of outputs in Riley is now deferred until the first call to HdxPrman_RenderPass::_Execute, and if the outputs change they get recreated in Riley,
allowing for dynamically editing the list of AOVs during an IPR session.
The Hydra display driver (framebuffer.cpp) had a hardcoded list of aov data that corresponded to the list declared to Riley,
so now that data reflects what is queried from the HdRenderPassAovBindingVector.

Technical details on some subtleties:
Prman only supports float3, float, and int outputs.  When a four component color is requested,
it is declared to riley as a color followed by a float alpha, and the framebuffer assumes this.
When hydra requests some other HdFormat that doesn't translate to float3, float, or int,
a conversion takes place in the Blit method of renderBuffer.cpp

(Internal change: 2075290)