Hydra: Changing argument passed to HdPerfLog::AddResourceRegistry from shared_ptr to raw pointer since HdPerfLog should not take (shared) ownership of a resource registry.
This exposed a problem where the task context data was implicitily holding on to resources from a render delegate causing a crash because the HdEngine managing the task context data was destroyed after the render delegate. The crash occured for testUsdImagingGLDomeLight where the task context data was holding on to a HdStSimpleLightingShader was holding on to various dome light textures.

To fix the problem, UsdImagingGLEngine explicitly deletes HdEngine before the render delegate.

(Internal change: 2077363)