Remove fall-through behavior for value clips
When value resolution encounters "gaps" in a clip set, where
some clips contain samples for a given attribute but others
do not, it will no longer fall-through and continue searching
weaker layers in the LIVRP(S) order for values.

This behavior had previously been required to avoid having to
inspect all clips to determine if an attribute's value was
intended to come from value clips. Instead, we now rely on
the clip manifest to determine if this is the case. An
attribute must exist in the manifest in order for value
resolution to look in value clips for a value. If it does and
a gap is encountered (i.e., an attribute's value is requested
at a time when the active clip does not contain samples for
that attribute), the default value specified in the manifest
will be used instead. If no default value is specified in
the manifest, the default value for the attribute type will
be used.

Note that USD will automatically generate an in-memory
manifest if one is not supplied.

The previous behavior allowed values from weaker layers to
shine through the gaps, which had several downsides:

- It was inconsistent with USD's typical layering behavior,
  where attribute values in stronger layers would completely
  override those in weaker layers. For example, if a user applied
  a clip set with gaps to override the animation on a referenced
  model, they would expect the final look to stay the same even
  if the overridden animation were changed. However, this would
  not be the case in the gaps due to the fall-through behavior.

- It meant that an attribute's value source was time dependent
  when clips were involved. This was complicated to explain and
  required special affordances in the public API,
  e.g. UsdAttribute::GetResolveInfo requiring a time, and the
  special UsdResolveInfoSourceIsTimeDependent source value.
  This time dependency also meant that UsdAttributeQuery could
  not accelerate queries to attributes that were affected by
  clips.

- The code to handle this behavior was complicated and buggy and
  would have required even more complexity to support future
  behavior changes for cross-clip interpolation.

Removing this behavior resolves these issues, for the most part.
UsdAttribute::GetResolveInfo still requires a time currently,
but the UsdResolveInfoSourceIsTimeDependent value is no longer
needed. Ultimately, this means a clip set is treated more akin
to a single layer whose time samples happen to be distributed
across multiple files.

(Internal change: 2077897)