Make bool operator on HgiHandle explicit.
We do this to avoid using it the wrong way while we transition Storm to Hgi.
You could accidently give a HgiBufferHandle to an OpenGL call and it would convert the handle to a bool and then the bool to a GLuint and compile ok, but not work.

(Internal change: 2078040)