Add garbage collector to HgiGL.
This is helpful when recording (deferred) Cmds in Hgi.
Example:

BlitCmds cmds;
for (buffer : buffers) {
    cmds.CopyBufferGpuToGPU(buffer);
    DestroyBuffer(buffer)
}
SubmitCmds(cmds);

In the above example we ask buffer to be destroyed during the for-loop before 'cmds' that uses them is submitted / executed. This type of loop was previously valid since opengl calls were 'immediate'.
Now they are delayed until SubmitCmds.

HgiMetal support this type of DestroyBuffer pattern since it also use garbage collection.
So this brings HgiGL inline with those other backends.

We use thread_local to also make Destroy requests thread safe (and mostly lock free).
Although we do not yet use threading in Storm for this, Metal and Vulkan both support this too so we bring HgiGL inline with the other backends.

(Internal change: 2078254)