Fixing the following subroot reference with variant selection bug.
In the bug's example, we have a prim /Model with variant selection {v=x} inside a reference and a child prim /Model/Child that has a sub-root reference to /Ref/Child, and /Ref has variant selection {v=y} authored.

Composition winds up using the {v=x} selection authored on /Model when composing the sub-root reference to /Ref/Child. This is incorrect and we should be using the authored selection {v=y} instead. Variant selections on ancestral prims (/Model) shouldn't be applied to child prims (/Model/Child). Pcp tries to maintain the same variant selection for prims at the same depth of introduction, but shouldn't be doing this across reference boundaries.

(Internal change: 2083085)