[hdEmbree] Fix issues with changing reprs (specifically, refined <-> unrefined transitions) in embree meshes.
We were missing a call to detach the old mesh geometry from the mesh scene, before we deleted it and recreated it.  This was causing crashes in the embree BVH builder and other badness.

While tracking this down, I improved the HdEmbreeMesh destructor a bit, and made _EmbreeCullFaces more robust to bad pointers; plus made it work correctly for ray packets, if we enable those later.

(Internal change: 2083299)