Store HgiResourceBindings and HgiPipeline objects on HdSt resource registry.
This addresses a performance regression where we are very frequently running the smooth normals computation and the CPU overhead of creating the descriptors (and filling its vectors) was causing a noticeable performance regression.

By storing the resource bindings and pipeline objects in the registry, we create them less frequently.

(Internal change: 2083939)