HgiGL: When resolving the textures at the end of executing the submitted graphics cmds, use framebuffers from the framebuffer cache instead of creating one every frame.
This requires the framebuffer cache to store two framebuffers for each graphics cmds descriptor, one for the MSAA and one for the resolved textures.

This is addressing a performance regression that happened when switching draw targets to use render buffers/AOVs instead of GlfDrawTarget's and thus using Hgi to resolve MSAA textures as last step of executing the submitted graphics commands.

(Internal change: 2084123)