UsdCollectionAPI is no longer generated using 'isPrivateApply = true' The UsdCollectionAPI now has public Apply function that can be used to apply a colleciton to a prim. ApplyCollection has been marked as a deprecated function as Apply is now the preferred function.
Note that UsdCollectionAPI originally had the private _Apply and custom ApplyCollection because API schemas were unable to provide built-in properties and fallbacks, so the expansionRule attribute needed to be instantiated to provide a value. Since then, API schemas were updated to be part of a prim's definition, providing built-in properties, so the custom instantiation of expansionRule is no longer necessary.

(Internal change: 2084387)