[usdImagingGL] Fix draw mode adapter handling of transforms for instanced card prims.
The draw mode adapter was handling transforms from the UsdPrim it was given.  However, instanced cards (being a bit of an oddity) get the USD prim of the USD instance.  In this case, the draw mode adapter needs to ignore transforms, since they are already explicitly tracked in the instance adapter.  If the draw mode also applies them, the cards object will be double-transformed.

Fixes #1251

(Internal change: 2085412)