Cache shader constant values in HgiGL op and destroy after op executes.
This is required since the data for the constant values might be allocated on the stack and will no longer be available when the op executes.

Metal and Vulkan copy the provided constant data when the command buffer records the command.
So we simulate this copy step in HgiGL.
(It might mean two copies. Once in HgiGL and one by the OpenGL driver, but the data is small, so should be ok)

(Internal change: 2089759)