HdPrman: Fix UsdUVTexture wrap mode support
In RenderMan, a texture's wrap mode is specified in the
texture asset metadata (.tex), not the shading node (PxrTexture).
To support UsdUVTexture, which allows specifying the wrap mode
as a shader parameter, we now pass the wrap mode arguments
into the RtxGlfImage texture plugin.  The default, "useMetadata",
will continue to consult the image metadata, as before.

(Internal change: 2091604)