[usdImaging] Only apply cards to prims with GeomModelAPI schema applied.
Requested by the USD team: previously, we applied cards when the attributes we needed were present, but now we only apply cards when the cards prim has the GeomModel API schema applied.

The exception to this is the feature to draw unloaded prims as bounds, which works the same way as before.

(Internal change: 2091694)