[usdImagingGL] Switch cards UVs from face-varying to vertex interpolated.
Using vertex interpolation for cards UVs means we need to bump up the vertex count (though as it turns out, only for "box" mode!); but if we avoid face-varying, we don't need to run the geometry shader, and we should get better utilization of primvar sharing (which is not currently applied to fvar primvars).

This should decrease the draw time of cards prims.

(Internal change: 2091696)