Add TextureView concept to Hgi.
A TextureView aliases the data of another Texture.

For example:

We have a texture used for a domelight and it contains several mip levels.
We want to use a compute shader to compute unique data for each mipmap.

To support this in all Hgi APIs we can use a TextureView that aliases into one mipmap.
This TextureView is then used by the compute shader to write into the mipmap.

Contribution: Jason Fielder

(Internal change: 2093503)