- Bundled Roboto and Roboto Mono fonts -- and made it default for usdview - Removed usage of Gotham Rounded - Removes a performance hack to change smallFont to avoid QT trying to generate suitable sized fonts. This is no longer noticed with Roboto or with other fonts.
Note that the change for the menu height being set to 25 is a result of saving the mainWindow.ui through designer5. Apparently qtdesigner sets it to 25 irrespective of an explicit value being set.
Also note that this is benign because the sizePolicy is "Minimum", which means Qt will make the bar taller when it needs to due to font size

(Internal change: 2094146)
(Internal change: 2094182)
(Internal change: 2094508)