[HdSt] On first material sync, try to batch prims with identical textures together.
Our batching code was batching prims together based on the scenegraph-location of its texture references.  This is effective in cases like a draw target.  However, for Usd cards, each cards prim specifies the cards textures separately (due to how the schema is written), meaning usd cards prims were ending up in a batch of 1.  This was terrible for performance when there were lots of cards prims, since the batch resource binding was vastly eclipsing the rendering time of the six quads in each batch.

Now, on the first sync of a material, we try to aggressively batch prims together if their texture asset paths match.  On the second sync, we revert to batching by scenegraph location.  This is an attempt to balance the needs of cards prims (to be aggressively batched together by asset path), with the needs of animated textures (to change texture assets each frame without invalidating batches).

Long-term, the better fix for this is bindless textures.  This change does almost as well as bindless textures, though, for the test cases we've examined.

Note that we're changing batching behavior for textures coming from a material network, but this change shouldn't affect batching behavior due to volume resources or dome lights.

A few other bonuses:
- HdStMaterialParam has a slightly more exhaustive default constructor now.
- HdStMaterial had a bunch of code in sync to re-batch when material tag changed, or when shader code changed.  That code was behind an if statement that was in the wrong place, making it effectively dead code.  I've re-enabled that code (since I needed to add another condition to the rebatching logic).

A final note: I was surprised to see multiple USD cards prims with the same textures, since I would have thought native instancing would collapse them all.  My suspicion is that this is happening because of the way point instancers break native instancing, though that's not 100% confirmed.

(Internal change: 2097606)