Call Hgi StartFrame and EndFrame to wrap one HdEngine::Execute.
We may place these calls somewhere else in the future, but we need to at least make sure Hgi garbage collection runs each frame or we would be leaking GPU memory quickly when a path tracer creates new images each frame inside AovInputTask and the old ones never get destroyed on the GPU.

The ideal scenario is to call Hgi::StartFrame at the start of HdEngine::Execute and Hgi::EndFrame at the end of HdEngine::Execute. But we do not want to link Hd to Hgi, so we use the first task's Prepare to StartFrame and the last task's Execute to EndFrame.

(Internal change: 2097667)