Prevent infinite loop in material processing in HdPrman
When there is a cycle in the network the recursive _ConvertNodes can get into an infinite loop. The previous checks for previously visited nodes did not catch cycles, since visited nodes are only added to the result vector at the bottom of the function, but the recursive part is at the top of the function.

This new version introduces an explicit set to check for visited SdfPaths, which is updated at the top of the file.

(Internal change: 2098774)