[hdx] Use the viewport as-is when AOVs are not used. Recently, we reworked the way camera masking in usdview was expressed. The changes to HdxTaskController were: - to use the viewport (width, height) as the dimensions of the AOVs. - suppress the viewport offset to the render task(s) and instead use it only during the compositing step.
For applications that use UsdImagingGLEngine to render into the bound GL framebuffer with Storm, this results in incorrectly translated (in screen space) geometry from the USD stage.
This change fixes this issue by using the render viewport as-is when AOVs aren't used.

Note: In the future, we expect to improve the UsdImagingGLEngine API to better communicate the AOV dimensions and the subRect (or fill region) on the bound framebuffer (if any). The current fix is meant to be a temporary bandaid.

(Internal change: 2098968)