Switch (most of) the GPU frustum culling code from raw gl to Hgi.
'Binder' is under the hood still using raw gl, so that will come in the next change where we switch that code to HgiResourceBindings. Left a few todo's to mark those areas.

This builds on top of the previous change that makes sure the barriers are set in HgiGL::SubmitCmds.

(Internal change: 2099824)