Change generated paths of USD native prototypes from "/__Master_<X>" to "/__Prototype_<X>".
Also change the path to flattened prototype prims generated
by UsdStage::Flatten from "/Flattened_Master_<X>" to
"/Flattened_Prototype_<X>".

Update tests and scattered comments to accommodate this.

Note that it has always been and continues to be the case
that downstream client code should never rely on the
format of these paths, as they are generated and subject
to change at any time.

(Internal change: 2100519)