[HdSt] Refactor instancer to have a sync phase, although the sync phase is currently called from rprims.
This is an attempt to clean up the relationship between instancers and rprims.  PopulateDrawItem is split into a function in rprimUtils; the instancer's "Sync Primvars" method updates the bar but lets rprims individually set those bars in their draw items; and instance index bar ownership is moved from instancer to rprim, which saves a bit of memory and a mutex hit.

Aside from being a cleanup, this is also meant to set the stage for syncing instancer topology.

(Internal change: 2118255)