Fix GPU memory leak of certain buffers never getting garbage collected.
When the artist closes a stage, we destroy representation of an imagable prim, which destroys the render index, delegate and Hgi.
The OitTask releases its shared_ptr to the oit buffers.
So in theory the next time garbage collection runs, you expect BufferArrayRegistry to free those GPU resources.

But because we create a NEW Hgi for the next stage, we also get a new ResourceRegistry and a new BufferArrayRegistry.

The old registry is still sitting in the 'resourceRegistry-Hgi' map, but no one is using it anymore, so no garbage collection is run on it. The OIT buffers GPU resources are never freed.

By calling GarbageCollect when the registry is destroyed, we make sure to do one more clearnup round on the registry.

Long term, we can hopefully get back to one ResourceRegistry and one Hgi for the entire application as it would be a lot less confusing.

(Internal change: 2118800)