Cache root layer stack identifier in PcpCache
PcpCache now holds on to the PcpLayerStackIdentifier
provided to its c'tor and returns a const reference
in PcpCache::GetLayerStackIdentifier instead of
recreating the object on each call.

This fixes performance regressions from change 2120435.
That change made copying an ArResolverContext more
expensive since doing so requires copying a std::vector
instead of a single shared_ptr.

In particular, GetLayerStackIdentifier is called multiple
times when retrieving targets and connections in Usd.
The cost of recreating the object quickly adds up,
especially when imaging large stages. With this fix,
performance in those cases should be better than it
was prior to the causing change.

(Internal change: 2121247)