Hydra Camera Schema (follow-up change to 2124799):
1. HdRenderPassState can compute the projection matrix when using physycally based attributes are given.
2. HdRenderPassState uses the camera's transform if no view matrix is given.

Note that there is a subtle change in behavior in that usdImaging was constructing a GfFrustum from the transform to then compute the view matrix. This would implicitly orthonormalize the matrix and loose scaling/shear. With this change, the matrix is no longer orthonormalized and scaling/shear makes it across to the render delegate. Similarly, an orientation reversing transform is no longer flipped to an orientation preserving one.

We currently have two ways of how a scene delegate is passing cameras to the render delegate. The scene delegate can either specify OpenGL-style view/projection matrices or specify physically based attributes such as horizontal aperture and focalLength together with the camera transform (that is the view matrix - at least if the transform has no scale/shear). Some render delegates only support the former (e.g., Storm) while others only support the latter (e.g., hdPrman). usdImaging was implementing both schemes so that it would work with either render delegate.

Longterm, we want to move away from the view/projection matrices towards physically based attributes. This change is a step in that direction.

(Internal change: 2124805)