Extend UsdShadeConnectableAPI to handle multiple connected sources
The main changes are all on the UsdShadeConnectableAPI, where multi-connection aware APIs are introduced. Sometimes as new methods and and sometimes by adding new parameters with default values to existing methods. Old methods have been marked as \deprecated, but the API is still compatible with the previous version (requires a recompile). The main transitions are:

* ConnectToSource with many parameters -> ConnectToSource using UsdShadeConnectionSourceInfo struct and ConnectionModification parameter
* Added SetConnectedSources
* GetConnectedSource deprecated in favor of GetConnectedSources
* GetRawConnectedSourcePaths deprecated in favor of GetConnectedSources (or GetAttr().GetConnections())
* DisconnectSource can disconnect a single targeted source
* ClearSource -> ClearSources

The UsdShadeInput and UsdShadeOutput classes forward the UsdShadeConnectableAPI for convenience and they have been updated in an analogous way.

New types introduced:
* enum UsdShadeConnectionModification
* struct UsdShadeConnectionSourceInfo

The GetValueProducingAttribute method on UsdShadeInput has been generalized and put into utils as GetValueProducingAttributes, with an overload for Output attributes. The new function is available on UsdShadeInput and also on UsdShadeNodeGraph as a replacement for ComputeOutputSource, which did something very similar, but less general and without a protection against infinite cycles.

All call sites in this library and the tests of deprecated functions have been updated to call the newer versions. The Connectability test has been extended to execise the multi-connection APIs

(Internal change: 2125207)
(Internal change: 2125275)
(Internal change: 2125339)