Add SdfLayer::GetResolvedPath
This is equivalent to SdfLayer::GetRealPath but returns
an ArResolvedPath, which is helpful for clients since
the ArResolver interface for Ar 2.0 uses this object
in many places. Also, its naming does not include the
filesystem-specific connotations that GetRealPath does.

To accommodate this, some SdfLayer internals were
updated to operate in terms of an ArResolvedPath.

(Internal change: 2127309)