[ar 2.0] Remove fileVersion from SdfLayer::UpdateAssetInfo
This parameter was legacy code that was added to work
around a Pixar-specific asset system issue long before
USD's open source release. Removing it is a step toward
removing "file version" as a first-class concept in Ar
and Sdf and will allow us to remove the awkward
ArResolver::UpdateAssetInfo function.

(Internal change: 2127528)