Adds support for varying interpolation of any basis curves primvar, not just widths and normals, by adding a templated interpolator computation class. Includes more specific checks on lengths of vertex or vaying authored primvars.
To allow for different behaviors of vertex and varying primvars, we now have a varying BAR, currently only used by basis curves. Varying primvars get accessed via buffer array in the vertex shader but are output as VS outputs and treated similarly to vertex primvars for the rest of the shader pipeline.

Fixes #1308

(Internal change: 2129963)