Storm texture system: simplifying logic what mips to load from a file for UV textures.
There are some subtle changes in behavior when the sizes of the authored mips aren't valid in that the next mip isn't half the width and height (rounded down) from the previous one. Whereas previously, all authored mips would be discarded if any one wouldn't have half the resolution of the previous one, we now have a best effort approach. That is all authored mips are considered to find one that fits into the given target memory. And once suc a mip has been found, we use as many of the following lower resolutions as we can, i.e., all mips until we find one that is not half the resolution of the previous one.

(Internal change: 2130838)