Fixes for testRmanOslParser
Set RMAN_SHADERPATH to the location of the OSL shaders used
by the test. These are typically found automatically by the
rmanDiscovery plugin by looking relative to the hdPrmanLoader
plugin, which is installed in <inst>/plugin. However, in the
monolithic build this plugin is embedded in the single shared
library, which is installed in <inst>/lib. This breaks the
automatic discovery, so we need to hard-code the path to
the shaders instead.

Also remove unnecessary quotes from env setting; these quotes
were being passed all the way through to rmanDiscovery causing
incorrect search paths.

Fixes #1396

(Internal change: 2131436)