Finishing work to support fieldDataType and vectorDataRoleHint for UsdVol prims in HdPrman
The mapping of the UsdVol type tokens is volume ImplicitField plugin specific, which is why we push the step to declare the fields of a volume as primvars into the specific emitter functions. We update the OpenVDB and Field3D implementations to support this. A DeclareFieldPrimvar helper function has been added to HdPrman_Volume.

Both versions have their own functions to map from the UsdVol settings to the appropriate primvar types. The OpenVDB mapping is more general than the current impl_openvdb implementation. This way we only need to update the ImplicitField plugin once the support has been added. There are comments in the code indicating that the current plugin in not supporting all types. The plugin will report any unsupported types.

(Internal change: 2145289)