Performance improvement for terminal resolution in UsdShade
The new-ish GetValueProducingAttributes functions are used by the code in UsdShadeMaterial::_ComputeNamedOutputSources. This has two issues: conceptually this could report attributes with authored values, which are not suitable for a terminal (surface, displacement, volume) connections. Secondly the code has slowed down compared to the previous specialized code, since the more general GetValueProducingAttributes was checking for authored values for every attribute it visited.

To remedy both issues we introduce a `shaderOutputOnly` flag on the GetValueProducingAttributes function, this allows us to get only the type of connections we want and avoid the cost of the extra checks. The only place that current uses the new version is in the material code that has the issues in the first place.

Also:
* The tests for GetValueProducingAttributes have been expanded to test the new version of the API calls.
* The documentation on "Connection Resolution Utilities" has been expanded as well as the definition of Containers
* Fixed various little typos in the documentation

(Internal change: 2147279)