Remove check for a python instance in your path
The build scripts will use the python interpreter that was used to
invoke them, rather than looking for an executable named python in the
path. Relying on python's name isn't very reliable currently. In Linux
python 3 builds the executable will be named python3, for instance.
Also, in some containerized environments python may be present with a
different name, or not be available in the path.