[hd, hdSt, usdImaging] Don't compute material network primvars in usdImaging; fix up the hdSt computation of material network primvars.
usdImaging primvar filtering was exclusively used by hdSt.  With the addition of hdSt primvar filtering (at the Sync level, dropping unreferenced primvars), and the removal of the primvar data from the value cache, this wasn't gaining us anything, and was holding us back architecturally; so this code is being removed as a simplification.

hdSt primvar filtering was already checking sdr for referenced primvars, but needed an extra clause to correctly process texcoord references.

(Internal change: 2149602)