arch: add ARCH_EXPORT_TYPE for type_info & vtable exports
For Linux and macOS builds, we need additional annotations to ensure
that type_info and vtables are exported for types when building with
hidden visibility.  Clang supports a type_visibility attribute that
controls visibility for exactly the entities that we care about.  gcc
does not provide a similar attribute so we have to export the type and
all its members, which is not ideal.

(Internal change: 2154168)