[usdImaging] Material binding improvements for point instancer prototypes.
This changelist should fix direct material binding for point instancer prototypes, with the important caveats that "strongerThanDescendants" and anything collection-based might not be honored, particularly around the intersection of nested point and USD-native instancers.

Previously, point instancers were not involved in material binding resolution, and delegated that to the prototype; however, if the prototype is inside a USD prototype root, it can't resolve bindings at the USD instance level.  The fix is to give a second chance at material binding resolution to the point instancer, which builds the proxy prim path to the prototype and runs the material binding query on that proxy path.

This should specifically fix cases where a material is bound upstream of a USD native instance that contains a point instancer prototype.

Fixes #1449

(Internal change: 2154991)