[usdImaging] When checking primvar existence, use Get() instead of ComputeFlattened().
In _ComputeAndMergePrimvar, we need to know primvar existence at time T, to know whether to include it in the descriptor list, but we don't actually use the value.  Previously we were calling ComputeFlattened; possibly a holdover from when this function actually prepared the buffer for imaging.  We can't use HasValue, since it doesn't capture time varying blocks; so we use Get.  Get should be fast as long as we don't touch the data; ComputeFlattened reduces to Get unless a primvar is indexed, in which case it pages in the data and does the actual flattening, which we don't want in this function.

(Internal change: 2155958)