Command line utility usgendynamicschemafromsdr
Uses a config file with sdrNodes identifier and sourceType and
a list of layers (to be sublayered) produces a schema.usda with
prim and property spec appropriately populated from the sdrNodes,
and isDynamic metadata set.

Once schema.usda is updated, this scripts runs usdGenSchema to
create generatedSchema.usda and plugInfo.json.

(Internal change: 2159923)