Fix specifier inconsistencies for the pseudo-root and instancing prototype prims.
These prims are "defined" but did not report having a
defining specifier which is inconsistent with the
documented behavior. This has been fixed. In addition,
the pseudo-root now reports its specifier as "def"
instead of the previous fallback of "over". This was
already the behavior for instancing prototypes.

(Internal change: 2160114)