Adds support for refinement of face-varying primvars.
Refinement of face-varying primvars requires knowing the primvar's topology, and different primvars on the same mesh can use the same or different topologies. This change adds a helper class _FvarTopologyTracker to keep track of the face-varying topologies and which face-varying primvar is associated with each of the topologies.  The gathering of face-varying topologies and the creation of face-varying index builder computations happens in HdStMesh::PopulateTopology().

The geometric shader now holds an FvarPatchType (determined in HdStMesh::_UpdateDrawItemGeometricShader) that will be used during codegen in face-varying primvar interpolation.

ResourceBinder and Codegen now support the codegen'ing of face-varying indices and patch params when applicable. When creating the face-varying primvar metadata in ResourceBinder, we also add the channel number so the primvar will know which face-varying indices and patch params to use. The geometry shader accessor for face-varying primvars has been heavily modified to support the various fvar patch types.

(Internal change: 2162201)
(Internal change: 2162458)