Reorder members to optimize size of UsdResolveInfo
Before change 2161341, sizeof(UsdResolveInfo) was 80 bytes.
After that change, that size increased to 88 bytes, causing
a small but noticeable regression in some performance tests.

This change reorders the members in UsdResolveInfo from
largest to smallest to pack them in tighter. This reduces
the size of UsdResolveInfo back to 80 bytes.

(Internal change: 2164201)