[Hd,HdSt] Removed redundant ptexId field from patchParam
This is redundant in the current implementation. In all cases this
"ptexId" is equal to gl_PrimitiveID for each drawn quad. Eliminating
this saves some GPU memory and will help avoid confusion as
we move forward with porting this code to run on Metal, where
we will have to index quad faces differently.

(Internal change: 2166474)