Always detach referenced ranges when closing a .usdc file mapping
Referenced zero-copy ranges within a mapped .usdc file must be
"detached" when closing the mapping to ensure that VtArrays
using those ranges remain valid afterwards. This was not
happening when saving a .usdc file, which recreates the file
mapping in CrateFile::Packer::Close.

The call to _DetachReferencedRanges now occurs in the
FileMapping object's d'tor to ensure it happens whenever
a mapping is closed.

(Internal change: 2167494)