Use ArWritableAsset for writing .usdc files
The new testUsdFileFormats_asset test exposed a
pre-existing issue with saving .usdc files when
using the ArAsset-based codepath and an
ArFilesystemAsset to read .usdc files on Windows.
This codepath isn't typically used; it's meant
to handle non-file backed .usdc assets so we
hadn't seen it until now. The test is disabled
for now while we look into a fix.

(Internal change: 2168790)