Adding CanApplyAPI functions to UsdPrim that can be used to determine if an API schema is valid to apply to a particular UsdPrim before applying it.
In support of this API, this change also adds fields to the Type plugInfo metadata
for appied API schema types that allows the type to specify "can only apply to"
prim types. Specifying "can only apply to" types indicates to the schema registry
and the CanApplyAPI functions that the API schema is only valid when applied to
prims whose type is one of the specified types.

And specifically for multiple apply API schemas, additional plugInfo metadata
can added to specify 1) the "allowed instance names" which will indicate that
only the listed instance names are valid for the schema and 2) a dictionary of
per instance name metadata that can specify "can only apply to" types for
specific instances of the multiple apply schema.

This change only handles the USD support for utilizing this new plugInfo
metadata; a following change will address the creation of this metadata through
usdGenSchema.

(Internal change: 2170113)