[usdImaging] Fix resolution of symlinked UDIM textures
Symlinked UDIM textures would cause issues if the symlinked
files did not follow the UDIM naming conventions. This was
because the UDIM resolution code in UsdImaging would take
the realpath of the first tile found and use that to
establish the expected pattern for the remaining tiles
instead of using the symlink itself.

This change removes the symlink resolution for UDIM
textures.  This avoids the issue above and seems less
confusing, but this does remove potential deduplication if
two different UDIM texture paths were symlinks to the same
files. However, this deduplication was also buggy. If we had
two texture paths A.<UDIM>.exr and B.<UDIM>.exr and the
first tiles were symlinks to the same file but the remaining
tiles pointed to different files, the old code would have
treated A and B as the same texture.

Fixes #1329

(Internal change: 2170800)