[usdrecord] Fix for Qt window being displayed
Instead of doing a widget->show and then widget->hide, we can simply set the gl context to be current, which will NOT make the qt widget show up. Note that we simply need to record the framebuffer in an image and do not want to show the gl widget.
This change prevents the usdrecord window from being displayed for a split second.

Fixes #1521

(Internal change: 2170955)