[ar 2.0] Add ArResolver::CanWriteAssetToPath
This change renames ArResolver::CanWriteLayerToPath to be
more general and modifies it to use the public NVI pattern
like the other ArResolver methods.

CanWriteLayerToPath had previously been deprecated and was
slated for removal in favor of moving its functionality to
to ArResolver::OpenAssetForWrite. The idea was that this
would place the checks next to the code responsible for
opening an asset for writing and would reduce the size of
the ArResolver API.

However, this didn't account for the possibility that some
layer file formats would not be able to use OpenAssetForWrite;
for example, formats that use an external library with
their own file writing code. So, CanWriteLayerToPath was
reintroduced and updated to ensure that its restrictions
would apply uniformly to all layers.

(Internal change: 2172550)