Fix handling of short reads/writes in ArchPRead/ArchPWrite
The total number of bytes processed was tracked incorrectly,
which caused these functions to report that it read/wrote more
bytes than requested.

This showed up as errors when reading large text layers via
ArFilesystemAsset, since that code errored out if the number
of bytes read did not equal the number of bytes requested.

This change adds a test utility, testArchPRead, that will
use ArchPRead to read a given file and dump out the last
N bytes of the file.

(Internal change: 2177227)