[usdImaging] Fix crash when processing multiple native instancer edits in the same frame.
This is actually a rehash of github issue #1417, which was the same bug in PointInstancers rather than native USD instancing.  If we get multiple property edits to native instancers in the same frame, it's possible that the first will resync the instancer and erase instancer map data, and then the second property edit will be unable to find the instancer to update.  This is totally expected and fine, since the resync will repopulate everything correctly, but our TF_VERIFY statements were a bit too aggressive.

This change relaxes the TF_VERIFY calls in the edit codepath.  I also fixed up a few stray callers of the affected utility functions to do return value checking/TF_VERIFY where appropriate.

Fixes #1551

(Internal change: 2178348)