Refactoring UsdSchemaRegistry's FindAndBuildAllSchemaDefinitions in preparation for allowing single apply API schemas to have dependencies on other API schemas.
The maps of type to UsdPrimDefinion are now initially populated while the plugins for
generated schema layers are being processed. Then afterwards the prim definitions
themselves are populated from the already discovered prim spec paths. These prim
definitions are populated in the order of:
  1. All mulitple apply API schemas (don't depend on other schemas)
  2. All single apply API schemas (don't depend on other schemas, YET!)
  3. All concrete typed schemas (depends on all API schemas, but no other schemas depend on them)
This separation will help with the implementation of population of nested single apply
schemas coming up as only the population of the single apply API schemas will have to
be updates to handle possible dependencies on other single and multiple apply schemas.

(Internal change: 2178807)