Refactoring step to simplify and clean-up the _GetTypeAsSdfType function
This is in preparation of adding multiple versions of this function for the upcoming support for multiple USD encodings.

The only behavior change is, that before a dynamic array property of an asset identifier with arraySize=0 would have not been correctly identified as an array and would have returned SdfValueTypeNames->Asset. This is fixed with this change. No fallout is expected, since this is a practically non-existing combination.

(Internal change: 2178936)