[hd] Don't emulate dirty bits for non-scene dirty bits.
When scene index emulation is turned on, the (scene delegate, dirty bits) API is translated into (scene index, dirty locator) format and back.  Locators represent scene data that's dirty, but there are some dirty bits that don't correspond to any scene data (e.g. InitRepr, Varying), and instead are used for render index or render delegate internal state management.  If we call MarkRprimDirty with only one of these bits (for example, mark all rprims with InitRepr), we want to skip the locator translation for performance reasons.

This is reflective of the fact that the per-prim dirty tracking of scene data and per-prim tracking of dirty list state/follow-on computations should really be in separate systems.

(Internal change: 2183906)