[hd] Speed up emulation prim removal.
Mainly: HdRenderIndex::RemoveSubtree is dramatically more expensive than a single-prim removal, because of the involvement of HdPrimGather.  This change does a bit of extra bookkeeping in sceneIndexAdapterSceneDelegate to see if it can get away with calling single-prim removal, for a net win.

This change also removes a stray (unused) map lookup in RemoveRprim.

(Internal change: 2185102)