[hd] Emulation: speed up delegate sync; implement PostSyncCleanup().
This CL speeds up delegate sync by gathering per-prim delegates in a parallel for loop, and caching delegates frame-to-frame as long as no prims were added/removed.  This makes the assumption that any prim being emulated will not get a new scene delegate pointer without a corresponding PrimsAdded message, which is a safe assumption.

The cached delegates are also used to implement PostSyncCleanup(), which was previously unimplemented.

(Internal change: 2185315)