[hd] Enable scene index emulation.
Scene indices are a new internal representation of scene data for imaging, as described in the white paper "Moduler Scene Delegation and Manipulation in Hydra".  This change turns on the internal scene index representation in hydra by changing an environment variable default: HD_ENABLE_SCENE_INDEX_EMULATION = true.  Plugin APIs and application APIs are unchanged for now, so turning on scene indices shouldn't result in user-visible changes, but this lays the groundwork for future imaging features.

(Internal change: 2185534)