[hd] Speed up GetInstanceIndices in emulation.
GetInstanceIndices was implemented naively in emulation, construcing the whole instancer topology for each prototype.  For highly branched instancers, this can explode the runtime of this call.  The fix is simply to make the pull of instance index buffers lazy (i.e. push the scene delegate access one level down in the hierarchy).

(Internal change: 2186008)