[hd] GetInstanceIndices() speedup: cache the top level instancer topology datasource.
This change caches the Hd_InstancerTopologyDataSource, primarily so we can cache the result of GetInstancerPrototypes and the translation of said value to a VtArray.  Those two operations can end up being surprisingly expensive, since if we do them once per prototype rprim we end up with O(#prototypes^2) SdfPath copies and some of our scenes are very highly instanced.

Caching with an atomic guarantees we get some caching and answers are always correct, but if there's contention in _GetInstancerTopologyDataSource we can end up doing duplicate work.  In some unscientific testing, this version is a bit slower than synchronizing on a mutex (1.3x on one of our highly instanced sets); but it benefits from not having a std::mutex per prim, which would cost a lot of memory.  With this change, the runtime of GetInstanceIndices() is once again reasonable and no longer the tall nail in our traces.

(Internal change: 2186216)