[hd] Speed up UsdImagingDelegate::Sync() when some of the request paths are not present.
In scene index emulation, UsdImagingDelegate::Sync() can receive prim paths of other delegates.  If we discover these, we want to continue to the next prim path as soon as possible.  Previously, the first thing we did with these paths was to convert index -> cache path, and then check if the prim info was present.  However, if the prim info isn't present, the index->cache path map entry won't be present, and we'll do an SdfPath::ReplacePrefix before discovering that the primInfo map is empty.

Now, we directly look at the index -> cache path table, skipping the ReplacePrefix.  If we don't find the path translation entry, we know the prim info doesn't exist, since the two datastructures are maintained in parallel.

(Internal change: 2186829)