Fixed bug that could cause UsdNotice::ObjectsChanged notice emitted after a call to UsdStage::Reload to contain pointers to destroyed objects, leading to invalid memory accesses when calling certain methods on the notice.
Change 2175340 added a batching mechanism for processing
changes in UsdStage, primarily to allow stage changes that
occurred during a call to UsdStage::Reload to be combined
into a single round of notification. These changes could be
due to a resolver change, or layers being reloaded.

In the latter case, UsdStage::_HandleLayersDidChange would
save pointers to data in the SdfNotice::LayersDidChange
notice it received so they could be included in the
ObjectsChanged notice emitted in _ProcessPendingChanges.
However, _ProcessPendingChanges would be called *after* the
LayersDidChange notice went out of scope, leaving the
pointers in the ObjectsChanged notice dangling.

This fix makes UsdStage::_HandleLayersDidChange process and
send out the ObjectsChanged notice immediately to ensure the
LayersDidChange notice is alive when the ObjectsChanged
notice is sent. This should still result in a single round
of notifications if UsdStage::Reload causes resolver change
and layer reloads, since the layer reloads happen after
the resolver changes are processed.

Fixes #1605

(Internal change: 2186836)