[hd] Re-enable Scene Index Emulation.
Scene Index Emulation is a major new imaging feature.  It was originally turned on in CL 2185534, but due to some performance issues it was disabled in CL 2186347.  We've addressed many of the performance issues now, so we're re-enabling it.  We expect to continue working on performance after this CL, though.

(Internal change: 2187040)