hdPrman should use RtxHioImage for lightColorMap.
We want the ability to pass asset paths that refer to things other than
.tex files on disk to prman for the lightColorMap property on lights.
We can do this by using the RtxHioImage rtxplugin but there are three
complications:

1) We must distinguish (at least) .tex files. Users will
not be able to read them via HioImage so these asset paths must be
passed directly (without using rtxplugin).

2) Unlike other textures, lightColorMap textures should not be
flipped vertically.

3) Only image assets should use RtxHioImage.

This change changes hdPrman/light.cpp to use the RtxHioImage plugin
with flipping disabled for resolved asset paths that don't have a .tex
extension for the texture:file input.  It also changes material.cpp
to use the RtxHioImage plugin for resolved asset paths that don't have
a .tex extension, and are either image assets on disk or for a light's
texture:file input.  This is a heuristic: an unresolved asset path is
probably (?) a relative path for RMAN_TEXTUREPATH to find, and a
resolved asset path that has the "tex" extension could really only
be a .tex file on disk.
If the resolved asset is on disk then we can check to see if it's an
image.  If it's for a light's texture:file we know its an image and
it must not be flipped.

This change adds a new TfDebug code, HDPRMAN_IMAGE_ASSET_RESOLVE, that
prints resolved image asset paths.  It's now used for lightColorMap,
assets in materials, and material filters.

(Internal change: 2189040)