Fix regression introduced in change 2188569
Per comments in the file, we need to make sure that
_assetModificationTime is reset to an empty VtValue in
SdfLayer::SetIdentifier if the resolver returns an
invalid ArTimestamp since this indicates the layer
hasn't been serialized to its new location yet. Prior to
the above change, ArResolver would have just returned
an empty VtValue in this case.

(Internal change: 2189050)
(Internal change: 2189055)