[usdImaging] Fix a possible crash when invalidating items, then deleting them, then drawing.
UsdImagingDelegate uses _dirtyCachePaths to maintain a list of prims to call UpdateForTime on, based on recent invalidation.  When prims in the underlying stage are deleted, we don't prune this list (for efficiency of removes), which means that it's possible for _dirtyCachePaths to point to nonexistent prims.  Previously, this would trip an assert; this CL relaxes the assert.

(Internal change: 2189221)