UpdateSchemaWithSdrNode now adds shaderId attrSpec with default value corresponding to sdrNode identifier
- usdgenschemafromsdr is updated such that schemaConfig now provides a renderContext ("ri" for renderman nodes as an example), this is then passed down to UsdUtilsUpdateSchemaWithSdrNode which uses this information along with SdrShaderNode's context field to create an attribute spec for shaderId and assigns it the shader's identifier as a default value.
- cleaned usage of string constants in usdgenschemafromsdr

Note that we need to provide identifier (overrideIdentifier parameter) to be used for sdrNodes which are queried using explicit asset paths. This should only be used when clients know the identifier being passed is the true identifier which sdr Runtime will provide when querying using GetShaderNodeByNameAndType, etc.

This change also updates the usdRiPxr schemas by appropriately running the updated usdgenschemafromsdr on the "updated" rman schemaConfig.json files.
       - also updates these schema to reflect changes from usdLux sdr nodes only having core types properties and shadowAPI and shapingAPI properties.

(Internal change: 2190059)