Let hdSt draw BasisCurves that don't completely cover the supplied points or other primvars.
If the client supplies a primvar that doesn't meet one of the size
criteria, a fallback primvar value is used instead.  Previously,
the valid authored size options were:
a) Size 1
b) Size equal to the number of elements implied by the topology and
primvar class specifier (varying or vertex)
c) Size less than the number of vertices, provided there are curve
indices supplied

Now, we adjust (b) for curves with curveIndices. If a client supplies
indices (which specify a specific point index for each vert, in arbitrary
order, with multiples, and possibly not covering the whole points
array), primvars larger than numVerts are now permitted, and
the primvar is truncated according to the primvar size implied by
the topology and indices.

Varying primvars of unexpected size are still counted invalid.

Vertex primvars of unexpected size and lacking indices are also still
counted invalid, as this is more likely an authoring mistake than
an intentional sparsification as would be the case with indices.

This strictly increases the configurations that are valid.

This change does not tackle deindexing or compacting large
primvars which are only sparsely used, because at the moment
we have no mechanism in place for aliasing point IDs for drawn
points.

(Internal change: 2190663)