[hd, hdSt, usdImaging] Change the format of UsdImagingDelegate::Get("/PointInstancer", "rotate") to VtArray<GfQuath>, to match SamplePrimvar; change expected hydra quaternion layout for "rotate" from (r,i,j,k) to (i,j,k,r).
Quaternions in USD are stored as GfQuath arrays, with (i,j,k,r) component layout.  Before this change, UsdImagingDelegate::SamplePrimvar() would return an array of GfQuath for the "rotate" primvar on a point instancer; however, UsdImagingDelegate::Get() would convert the quaternions to GfVec4f format, with (r,i,j,k) component layout.  The main motivation for this change was to resolve the API incompatibility, by having Get() return an array of GfQuath as well.

In order to not need to rewrite the array in processing, this code also adds half vector and quaternion support to HdVtBufferSource and hdStorm; GfQuath is now interpreted as a 4-component vector of halfs-floats, and hydra can transport this data all the way to the GPU.  This saves us the array copy, and also saves us some GPU memory and bandwidth since the data is sent packed.

hdStorm half support is implemented with GL_HALF_FLOAT for vertex attributes.  For other binding points, like SSBOs, the half floats are sent as uint[] and unpacked by the generated HdGet_ functions.  Ext-computation output to half-float vectors is also supported.  hdStorm only supports half2 and half4; half1 and half3 are omitted since they are more complicated to index into.

(Internal change: 2192012)