Fix HdStBasisCurves handling of invisiblePoints indices when the points primvar is oversized.
Previously, if the points primvar had enough unreferenced points at the
front of the array, and those points were marked invisible, it would trigger
errors computing the point visibility bitmask, and the resulting points
display would be incorrect.

Now, we:

- Ensure the vis bitmask is wide enough to contain bits for every point, even
if the point is unreferenced, so that these points may be invised.  However
in the interest of saving memory, we still truncate the unreferenced end
of the points primvar if it's oversized.

- Remove an incorrect clamping of point IDs.  We were clamping point IDs
to be at most the number of verts, which is no longer correct since we
recently allowed points arrays to be arbitrarily padded compared to the
points referenced by verts.

(Internal change: 2192309)