[Hd,HdSt] Added support for triangulated quad geometry
A triangulated quad is a 6-index tuple primitive representing
two triangles (3 indices each) of a triangulated quad. The
representation takes advantage of Storm's index array tuple
encoding so even though each quad primitive consists of two
triangles, there is no other data expansion of topological
or primvar data.

This is an optional feature that is not yet enabled.

(Internal change: 2199149)