[HdSt] Support face varying evaluation w/o geometry shader
Updated Storm codeGen to allow face varying primvar accessors
to be emitted to either the GS or FS. With support for built-in
barycentric coordinates and triangulated coarse and refined quads
Storm can now interpolate face varying primvar directly in the
fragment shader without needing an active geometry shader.

This is significant for Storm running on Metal (which does not
support geometry shaders), and it can also be more efficient
when running on GL or Vulkan.

Contribution: Thor Hjalmarsson

(Internal change: 2200681)