Add purpose/rendertag "widget" for Hydra widget-style drawing in Storm.
This change adds a special "widget" tag intended for prims like manipulator
rings and jacks.  When a "widget" render pass providing depth and color input
AOVs renders this category of geometry using the included fallback "widget"
shader, the result is that it draws atop existing drawn prims and blended
semi-transparently if it would have been occluded.

The "widget" fallback material is exempted from being considered a "scene
material" for the purposes of toggling possibly-costly scene materials on and
off, the idea being that these widgets are likely built into the host app and
should always shade as intended.

(Internal change: 2201266)