[Hgi] Added HgiVertexBufferStepFunction
Hydra Storm encodes drawing coordinate data in dispatch buffers
as vertex attribute data with a per drawing command step rate.

This adds vertexStepFunction to HgiVertexBufferDesc with enum values
to allow specification of HgiVertexBufferStepFunctionPerDrawCommand,
along with an implementation for HgiGL.

This is supported for Vulkan and GL multi-draw indirect buffers via
vertex attribute divisors, and will be implemented for the HgiMetal
implementation of DrawIndirect and DrawIndexedIndirect.

(Internal change: 2206477)