Add screen-space curve width controls and optional point scaling to BasisCurves.
The new screenSpaceWidths primvar toggles the interpretation of widths to be
in screen-space pixels.  We expect this to be useful for implementing guides
or other UI elements drawn with BasisCurves.  The pointsSizeScale primvar
similarly is intended to give clients a way to emphasize or supress certain
points by scaling their default size.

The new minScreenSpaceWidth primvar gives a minimum screen space width
in pixels for BasisCurves when rendered as tubes or camera-facing ribbons.
We expect this to be useful for preventing thin curves such as hair from
undesirably aliasing when their screen space width would otherwise dip
below one pixel.

(Internal change: 2206713)