[Hd] Slightly improved HdMeshUtil Patch Param documentation
The patch param computed when triangulating or quadrangulating a
mesh topology is encoded such that it can be decoded by compute
kernels and shaders running on the GPU. Updated the documentation
slightly to make this more clear.

Looks like there are additional opportunities to improve the doxygen
output here, but limiting the scope of this fix for now.

Fixes #1720

(Internal change: 2207820)