[hd] Add API to insert additional scene indices into the render index.
This adds an add/remove sub-scene API that is a rough match to the prim add/remove.  For the moment, the render index is maintaining the correspondence between scene prims and render prims via:
1.) Translating InsertRprim requests into scene index entries in HdLegacyPrimSceneIndex.
2.) Translating prims in the legacy prim scene index into render delegate objects.

Eventually we want to deprecate InsertRprim, so that scene prims show up in scene index format natively.  This API enables that, by letting us merge data between the emulation scene and any scene indices the app chooses to add.  Since the new scene index prims are indistinguishable from legacy prim scene index prims, SyncAll, HdSceneIndexAdapterSceneDelegate, and the render delegate don't need to change at all to consume direct scene index data.

Note that this API needs to be in the render index currently, but long term it would be better to split out the merging of input scenes and the storage/update logic of rendering objects.

Also note that these new functions are no-ops if scene index emulation is disabled.

(Internal change: 2208175)