[usdImaging] Add datasource handlers for basic prim data.
DataSourcePrim is a base class for other UsdPrim-based objects, and it's also the data source we create for untyped nodes.  Since we flatten inherited attributes downstream of the scene index actually reading the USD stage, we need to check every prim in the scene for potentially inheritable attributes.  Unfortunately, we can't do this by schema type and API types alone, but DataSourcePrim will aggressively early out based on IsA<> and HasAPI<> checks where applicable.

(Internal change: 2209419)