[hd] Add a default constructor to HdRenderBufferDescriptor.
It turns out HdSceneIndexAdapterSceneDelegate returns a default constructed render buffer descriptor when called on a prim that's not a render buffer; we need to add a default constructor so we don't just get uninitialized values.  This wasn't happening in practice, but the windows build helpfully warned about it.  Adding constructors to this struct matches what we do with other similar scene delegate structs.

(Internal change: 2211287)