[HdSt] Distinguish between displacement shader and GS source
Although Storm GL currently executes the displacement shader from
the geometry shader, the shader itself is not specific to the geometry
shader. For Metal (and potentially for Vulkan and GL) we'd like to be
able to execute the displacement shader from the tessellation stage.

This change removes a few more of the code sites where the
displacement shader was assumed to be in the geometry shader
source.

A follow up change could further clean up references to the
fragment and geometry shader stages from material network
shader processing.

Contribution: Thor Hjalmarsson

(Internal change: 2213140)