Remove SDR_DEFAULT_VALUE_AS_SDF_DEFAULT_VALUE environment variable
History:
We never had a way to request default values for a sdrProperty's SDR Type and as an SDF Type.
With change 2189531, we introduced a new function GetDefaultValueAsSdfType which corresponds
to the default value for the type returned by GetTypeAsSdfType. This GetDefaultValueAsSdfType was
the intended usage of GetDefault. Now with the introduction of this new method, GetDefaultValue's
implementation was modified to return the default value corresponding to the type associated with
Sdr (Sdr types are a subset of SdfTypes) returned by SdrShaderProperty::GetType().

The environment variable was introduced and was set to True, so as to have minimal impact from
the change (available in USD release 21.11). In USD release 22.03 this environment variable was
set to False, which means GetDefaultValue will ONLY return default value associated with the Sdr
type returned by SdrShaderProperty::GetType().

And now with this change, subsequent USD Release will be removing this environment variable all
together to fully establish the intended behavior.

(Internal change: 2213143)