As described in the docs (https://graphics.pixar.com/usd/release/api/usd_shade_page_front.html#UsdShadeConnectability)
A single Applied API schema can also register a connectableAPIBehavior and hence the assertion in the templated HasConnectableAPI
to check for the type being a UsdTyped schema is incorrected. With this change the assertion has been updated and requires the type
being queried is of type UsdSchemaBase (making sure a call to HasConnectableAPI with ANY other type, will return a false).

This also brings the python API and C++ API for HasConnectableAPI in a more consistent state.

(Internal change: 2215849)