Allow authoring to non-local opinions via instance proxies
UsdStage previously errored out when trying to author changes
via instance proxies because overrides on children of
prototypes (which are represented by instance proxies) are
ignored. However, this also restricted changes to non-local
opinions which are valid. For example, given instance prim
/A that inherits from /Class, opinions on /A/Child would
be ignored, but opinions on /Class/Child would not.

This change enables these non-local edits by checking
the stage's edit target when validating whether an edit
should be allowed.

(Internal change: 2217496)