Disable MaterialX scene index plugin when MaterialX support is not enabled.
This change just elides functions and headers that
aren't present when MaterialX support is disabled.
The MaterialX scene index plugin is still present,
but becomes a no-op in this case. Ideally the plugin
itself wouldn't even be built, but it's listed in
the plugInfo.json file and there's no easy way to
conditionally remove entries in that file.

(Internal change: 2217648)