[hdx, hdSt] Assets drawing black
Initializing the HasIntensity value to false when area lights
are invisible so we can ignore them in the simpleLightTask.
With this, disabling SceneLights will only add the built-in
lights to the context. Also adding a warning when we
surpass the maxLightsUsed.

Adjusting the order in which the lights are added to the
context so that the dome/camera(simple) lights are first.
This way even if we have more than 16 lights visible or
enabled, the first ones will be the builtin lights.

(Internal change: 2218061)