[hdPrman] Fix a hang in StopRender() when blocking = true.
Stopping riley in a blocking way is a little complicated, because we need to shut down both riley and the render thread, and riley can't look at the render thread's stop bit.  The previous approach calls riley->Stop() repeatedly until either we've successfully cancelled riley or the render thread callback exits; this is mainly to catch the case where you call StopRender() after the render thread is in the rendering callback but before it's called into riley.  However, this fails if we successfully cancel riley and then fail to exit the render thread for some reason.

The new approach loops (with a sleep) while the render thread is running and repeatedly calls riley->Stop(), ignoring the return value.  This results in more calls into riley->Stop() (although after riley is successfully stopped, these should be cheap), but in trade we no longer sit around waiting for the render thread to exit when riley is still running, as was happening before.  This specifically handles the cases of: the render thread calling back into riley for some reason, and riley->Stop()'s return value being untrustworthy.  It also means that if we get a hang, it should be much clearer now which part of the code is hanging.  (However, this fixes the hang we had a repro for).

Note that this hang wasn't a true deadlock, because even if you fail to stop riley, the render will eventually exit; but of course the distinction is lost if you're using an interactive workflow.

(Internal change: 2218106)
(Internal change: 2218864)