Add materialIsFinal to let certain drawing be immune to material overrides.
Previously, HdSt_MaterialNetworkShader had an IsSceneMaterial API that was used to
exempt storm fallback materials from
being pruned when the render pass state GetUseSceneMaterials API returned false.
The existing API was not exposed outside storm and thus was not available
for the host app to use to tag materials it supplied as "not scene" in order
to avoid pruning.

Now, we remove the IsSceneMaterial API and replace it with a
MaterialIsFinal API found on HdDisplayStyle, where the host app can set it
on a per-rprim basis.  This permits the host app to exempt various
non-scene geometry from having its materials pruned when the user
disables scene materials.

(Internal change: 2219406)