[HdSt,Hgi*] Added support for drawing tessellated patches with Metal
Added support to HdSt_PipelineDrawBatch for compiling and executing
drawing commands which use Metal tessellation. There is some added
complexity here because the render command encoding for drawing
patches with Metal is different than the encoding for other kinds
of primitives.

The initial use case for this is to use Metal tessellation for
Storm geometric primitives which can be used in place of geometry
shaders.

Because Metal does not support a multi-draw encoding for patches
we also extend the vertex buffer step function system in
HgiMetalGraphicsCmds to handle the base vertex offset for patches.
(See hgiMetal/graphicsCmds.h for details).

This change also updates the names of drawing parameters used in
Hgi*GraphicsCmds and HdSt_PipelineDrawBatch to be more consistent.

Contribution: Thor Hjalmarsson, David G Yu

(Internal change: 2221205)